Reorganize crategen to support multiple commands
