--- service_crategen/src/codegen/mod.rs
@@ -1,4 +1,0 @@
-pub mod botocore;
-pub mod generator;
-mod serialization;
-mod util;
