+++ service_crategen/src/commands/generate/codegen/rest_response_parser.rs
@@ -1,7 +1,7 @@
 use inflector::Inflector;
 
 use ::Service;
-use codegen::botocore::{Operation, Shape, ShapeType, Member};
+use botocore::{Operation, Shape, ShapeType, Member};
 
 // Rest Response Parser
 //
