+++ service_crategen/src/commands/generate/codegen/xml_payload_parser.rs
@@ -1,7 +1,7 @@
 use inflector::Inflector;
 
 use ::Service;
-use codegen::botocore::{Member, Operation, Shape, ShapeType};
+use botocore::{Member, Operation, Shape, ShapeType};
 use super::{generate_field_name, mutate_type_name};
 
 pub fn generate_struct_attributes(_deserialized: bool) -> String {
