+++ service_crategen/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod check;
+pub mod generate;
